fix OLPC #1015 - Use new colorchooser
In the Tools tab this patch changes the Tool Color button to Stroke
Color, and uses the new colour chooser.  In the Shapes tab, this patch
changes the Fill Color and the Stroke Color buttons, and uses the new
colour chooser.

Each of the colour choices made are effective for the next drawing
operation.

Tested-by: James Cameron <quozl@laptop.org>
Signed-off-by: James Cameron <quozl@laptop.org>