fix #1771 - paint overwrites file type instead of creating new file
Paint only writes type image/png.  It can load other types, but
overwrites as image/png.  This patch detects reading a type that is not
image/png, and forgets the association between the image being edited
and the journal entry that was read, so that the original journal entry
is not overwritten.

http://bugs.sugarlabs.org/ticket/1771