Initial phase of clipboard operations ported to cairo
Copy, paste and move operations are implemented.
NOTES:
* Change size of selection is not implemented in this patch.
* A refactoring was done to do the code easier to understand.
Previously, the Area.selmove was used to check if a selection is
in progress. Now only to check if a area is being moved.
A selection is detected using Area.is_selected().
The propery selmove now is private _selmove.
The selected area now should be obtained using Area.get_selection_bounds(),
and not using Area.orig_x, Area.orig_y and the selection size.
This properties should be removed after finish refactoring.
The property Area.sel_get_out was removed.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>