Use a similar surface to improve performance
As the surface is created in draw callback, we need delay some operations
until the surface is created. The undo keep is delayed until finish draw,
then enable_undo() now only enable a flag, and a new method keep_undo()
do the real task.
This patch breaks rotation and mirror, will be adressed in another patch.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>