Fix pick color math - SL #4333
The value obtained by read the pick color tool was wrong,
because the range is 0 to 255, and we use it between 0 and 65535
A similar problem is described here [1]

Aditionally, this patch set the right mouse cursor
for the case where the bucket do not fill anything,
because the selected color is the same to the color in the
point to fill. Previously the clock cursor was not removed,
and confused the user.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>

[1] http://old.nabble.com/-cairo--Getting-wrong-cairo-pixel-color.-td31256086.html