Avoid execute tool_end before the last tool_move - OLPC #12405
We are using a timeout in the mousemove events to avoid flood
with more events than the application can draw, but this can produce
get a move event after the method to finish the draw was executed.
As they draw in different canvas, the last part of the draw is lost.
This patch add a timeout a little bigger to the toolend call,
to avoid this situation.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>