Use GdkPixbuf to open files - SL #4418
Instead of use a cairo.ImageSurface, use a GdkPixbuf to load
files and paint it over the ImageSurface. In this way we can open
more file types than png, like in previous Paint versions.
This solves opening a clipboard object, like in the ticket SL #4418
Also the cairo developers said the ImageSurface.create_from_png method
is not designed to use in production but only to help in testing.

This patch is a simpler version based in the work of Manuel Kauffman.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>