Load stamps from journal and remember it.
This patch add Load from journal icon in the Load stamp
icon view. Just click it and object chooser will appear,
choose the image, and It will set as stamp.

Also, saves the ids of the images and add the image to the store.
When you relaunch the activity, if the object exists, its added
to the store