Fix text tool borrowing window underneath
When the text tool is used on Ubuntu 16.04 KVM, instead of the text
entered the area drawn uses pixels from the application window
underneath the activity window, which is usually the Sugar Home View.

A workaround in this commit is to repeat the fetch of window pixels.

The underlying problem is possibly a race condition exposed by the more
modern GTK+ toolkit and X server.

Does not occur on Ubuntu 16.04 on physical hardware.

Does not occur on Fedora 18.