Preallocate Area widget
On Fedora 27, recent Debian, and Ubuntu 18.04, the Area widget was not
seen.  The widget was default allocation 1x1, a single pixel, and not
easy to see.

Workaround was to show a toolbar menu, which causes a resize of the
canvas.

Existing code tries to set the Area widget to the correct size once the
activity widget is mapped and the scrolled window widget size is
allocated.

Before recent changes in GTK+, the second resize worked fine.  Now, the
second resize after the map is buffered somehow.

Rather than fixing GTK+, or making the activity incompatible with
previous versions of GTK+, we set an initial size of screen width and
screen height less grid cell size.

Fixes #27.